Removes production configs
