+++ substrate/frame/system/src/lib.rs
@@ -262,7 +262,7 @@ pub mod pallet {
 	/// Default implementations of [`DefaultConfig`], which can be used to implement [`Config`].
 	pub mod config_preludes {
 		use super::{inject_runtime_type, DefaultConfig};
-		use frame_support::{derive_impl, traits::Get};
+		use frame_support::traits::Get;
 
 		/// A predefined adapter that covers `BlockNumberFor<T>` for `Config::Block::BlockNumber` of
 		/// the types `u32`, `u64`, and `u128`.
@@ -290,7 +290,7 @@ pub mod pallet {
 			type Hash = sp_core::hash::H256;
 			type Hashing = sp_runtime::traits::BlakeTwo256;
 			type AccountId = u64;
-			type Lookup = sp_runtime::traits::IdentityLookup<u64>;
+			type Lookup = sp_runtime::traits::IdentityLookup<Self::AccountId>;
 			type MaxConsumers = frame_support::traits::ConstU32<16>;
 			type AccountData = ();
 			type OnNewAccount = ();
@@ -320,121 +320,6 @@ pub mod pallet {
 			type PostInherents = ();
 			type PostTransactions = ();
 		}
-
-		/// Default configurations of this pallet in a solochain environment.
-		///
-		/// ## Considerations:
-		///
-		/// By default, this type makes the following choices:
-		///
-		/// * Use a normal 32 byte account id, with a [`DefaultConfig::Lookup`] that implies no
-		///   'account-indexing' pallet is being used.
-		/// * Given that we don't know anything about the existence of a currency system in scope,
-		///   an [`DefaultConfig::AccountData`] is chosen that has no addition data. Overwrite this
-		///   if you use `pallet-balances` or similar.
-		/// * Make sure to overwrite [`DefaultConfig::Version`].
-		/// * 2s block time, and a default 5mb block size is used.
-		pub struct SolochainDefaultConfig;
-
-		#[frame_support::register_default_impl(SolochainDefaultConfig)]
-		impl DefaultConfig for SolochainDefaultConfig {
-			/// The default type for storing how many extrinsics an account has signed.
-			type Nonce = u32;
-
-			/// The default type for hashing blocks and tries.
-			type Hash = sp_core::hash::H256;
-
-			/// The default hashing algorithm used.
-			type Hashing = sp_runtime::traits::BlakeTwo256;
-
-			/// The default identifier used to distinguish between accounts.
-			type AccountId = sp_runtime::AccountId32;
-
-			/// The lookup mechanism to get account ID from whatever is passed in dispatchers.
-			type Lookup = sp_runtime::traits::AccountIdLookup<Self::AccountId, ()>;
-
-			/// The maximum number of consumers allowed on a single account. Using 128 as default.
-			type MaxConsumers = frame_support::traits::ConstU32<128>;
-
-			/// The default data to be stored in an account.
-			type AccountData = crate::AccountInfo<Self::Nonce, ()>;
-
-			/// What to do if a new account is created.
-			type OnNewAccount = ();
-
-			/// What to do if an account is fully reaped from the system.
-			type OnKilledAccount = ();
-
-			/// Weight information for the extrinsics of this pallet.
-			type SystemWeightInfo = ();
-
-			/// This is used as an identifier of the chain.
-			type SS58Prefix = ();
-
-			/// Version of the runtime.
-			type Version = ();
-
-			/// Block & extrinsics weights: base values and limits.
-			type BlockWeights = ();
-
-			/// The maximum length of a block (in bytes).
-			type BlockLength = ();
-
-			/// The weight of database operations that the runtime can invoke.
-			type DbWeight = ();
-
-			/// The ubiquitous event type injected by `construct_runtime!`.
-			#[inject_runtime_type]
-			type RuntimeEvent = ();
-
-			/// The ubiquitous origin type injected by `construct_runtime!`.
-			#[inject_runtime_type]
-			type RuntimeOrigin = ();
-
-			/// The aggregated dispatch type available for extrinsics, injected by
-			/// `construct_runtime!`.
-			#[inject_runtime_type]
-			type RuntimeCall = ();
-
-			/// The aggregated Task type, injected by `construct_runtime!`.
-			#[inject_runtime_type]
-			type RuntimeTask = ();
-
-			/// Converts a module to the index of the module, injected by `construct_runtime!`.
-			#[inject_runtime_type]
-			type PalletInfo = ();
-
-			/// The basic call filter to use in dispatchable. Supports everything as the default.
-			type BaseCallFilter = frame_support::traits::Everything;
-
-			/// Maximum number of block number to block hash mappings to keep (oldest pruned first).
-			/// Using 256 as default.
-			type BlockHashCount = TestBlockHashCount<frame_support::traits::ConstU32<256>>;
-
-			/// The set code logic, just the default since we're not a parachain.
-			type OnSetCode = ();
-			type SingleBlockMigrations = ();
-			type MultiBlockMigrator = ();
-			type PreInherents = ();
-			type PostInherents = ();
-			type PostTransactions = ();
-		}
-
-		/// Default configurations of this pallet in a relay-chain environment.
-		pub struct RelayChainDefaultConfig;
-
-		/// It currently uses the same configuration as `SolochainDefaultConfig`.
-		#[derive_impl(SolochainDefaultConfig as DefaultConfig, no_aggregated_types)]
-		#[frame_support::register_default_impl(RelayChainDefaultConfig)]
-		impl DefaultConfig for RelayChainDefaultConfig {}
-
-		/// Default configurations of this pallet in a parachain environment.
-		pub struct ParaChainDefaultConfig;
-
-		/// It currently uses the same configuration as `SolochainDefaultConfig`.
-		#[derive_impl(SolochainDefaultConfig as DefaultConfig, no_aggregated_types)]
-		#[frame_support::register_default_impl(ParaChainDefaultConfig)]
-		impl DefaultConfig for ParaChainDefaultConfig {}
 	}
 
 	/// System configuration trait. Implemented by runtime.
